+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/core/CliktCommand.kt
@@ -134,13 +134,7 @@ abstract class CliktCommand(
 
         val envval = readEnvvar(envvar) ?: return
 
-        val completion = when {
-            "fish" in envval -> CompletionGenerator.generateFishCompletion(command = this)
-            "zsh" in envval -> CompletionGenerator.generateZshCompletion(command = this)
-            else -> CompletionGenerator.generateBashCompletion(command = this)
-        }
-
-        throw PrintCompletionMessage(completion, forceUnixLineEndings = true)
+        CompletionGenerator.throwCompletionMessage(this, envval)
     }
 
     /**